services/controller_service/requirements.txt
kafka-python
requests
docker